// src/resolvers-types.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: any; output: any; }
};

export type Land = {
  __typename?: 'Land';
  abbr: Scalars['String']['output'];
  bday: Scalars['Date']['output'];
  name: Scalars['String']['output'];
};

export type Law = {
  __typename?: 'Law';
  bday: Scalars['Date']['output'];
  detail: LevelDetail;
  land: Land;
};

export type Level = {
  __typename?: 'Level';
  bday: Scalars['Date']['output'];
  dday?: Maybe<Scalars['Date']['output']>;
  details: Array<LevelDetail>;
};

export type LevelDetail = {
  __typename?: 'LevelDetail';
  index: Scalars['Int']['output'];
  name: Scalars['String']['output'];
  notes?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
};

export type Query = {
  __typename?: 'Query';
  laws?: Maybe<Array<Maybe<Law>>>;
};


export type QueryLawsArgs = {
  topic: Scalars['String']['input'];
};

export type Rule = {
  __typename?: 'Rule';
  bday: Scalars['Date']['output'];
  lands: Array<Scalars['String']['output']>;
  level?: Maybe<Scalars['Int']['output']>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;



/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  Date: ResolverTypeWrapper<Scalars['Date']['output']>;
  Int: ResolverTypeWrapper<Scalars['Int']['output']>;
  Land: ResolverTypeWrapper<LandModel>;
  Law: ResolverTypeWrapper<LawModel>;
  Level: ResolverTypeWrapper<LevelModel>;
  LevelDetail: ResolverTypeWrapper<LevelDetailModel>;
  Query: ResolverTypeWrapper<{}>;
  Rule: ResolverTypeWrapper<RuleModel>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Boolean: Scalars['Boolean']['output'];
  Date: Scalars['Date']['output'];
  Int: Scalars['Int']['output'];
  Land: LandModel;
  Law: LawModel;
  Level: LevelModel;
  LevelDetail: LevelDetailModel;
  Query: {};
  Rule: RuleModel;
  String: Scalars['String']['output'];
};

export interface DateScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['Date'], any> {
  name: 'Date';
}

export type LandResolvers<ContextType = DataContext, ParentType extends ResolversParentTypes['Land'] = ResolversParentTypes['Land']> = {
  abbr?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  bday?: Resolver<ResolversTypes['Date'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type LawResolvers<ContextType = DataContext, ParentType extends ResolversParentTypes['Law'] = ResolversParentTypes['Law']> = {
  bday?: Resolver<ResolversTypes['Date'], ParentType, ContextType>;
  detail?: Resolver<ResolversTypes['LevelDetail'], ParentType, ContextType>;
  land?: Resolver<ResolversTypes['Land'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type LevelResolvers<ContextType = DataContext, ParentType extends ResolversParentTypes['Level'] = ResolversParentTypes['Level']> = {
  bday?: Resolver<ResolversTypes['Date'], ParentType, ContextType>;
  dday?: Resolver<Maybe<ResolversTypes['Date']>, ParentType, ContextType>;
  details?: Resolver<Array<ResolversTypes['LevelDetail']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type LevelDetailResolvers<ContextType = DataContext, ParentType extends ResolversParentTypes['LevelDetail'] = ResolversParentTypes['LevelDetail']> = {
  index?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  notes?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type QueryResolvers<ContextType = DataContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  laws?: Resolver<Maybe<Array<Maybe<ResolversTypes['Law']>>>, ParentType, ContextType, RequireFields<QueryLawsArgs, 'topic'>>;
};

export type RuleResolvers<ContextType = DataContext, ParentType extends ResolversParentTypes['Rule'] = ResolversParentTypes['Rule']> = {
  bday?: Resolver<ResolversTypes['Date'], ParentType, ContextType>;
  lands?: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType>;
  level?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type Resolvers<ContextType = DataContext> = {
  Date?: GraphQLScalarType;
  Land?: LandResolvers<ContextType>;
  Law?: LawResolvers<ContextType>;
  Level?: LevelResolvers<ContextType>;
  LevelDetail?: LevelDetailResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  Rule?: RuleResolvers<ContextType>;
};